improve inbox rule handling
Instead of applying inbox rules based on the unread mail state in the
inbox folder, we introduce the new isInboxRuleApplied boolean flag on
the mail type. If a mail has been processed by the leader client, which
is checking for matching inbox rules, the isInboxRuleApplied flag is
set to true. If there is a matching rule the flag is updated through the
MoveMailService, if there is no matching rule, the flag is updated
using the IsInboxRuleAppliedStateService. Both requests are
throttled / debounced.

This new inbox rule handing solves the following two problems:
 - when clicking on a notification it could still happen,
   that sometimes the inbox rules where not applied
 - when the inbox folder had a lot of unread mails, the loading time did
   massively increase, since inbox rules were re-applied on every load

Co-authored-by: abp <abp@tutao.de>