[desktop] Change DesktopUtils#relaunch to DesktopUtils#exit
We can't trust the path in the environment variable even if it's correct
and factual, as we don't know if the file was modified. We also do not
want to spend a lot of time writing code to verify it right now.

To keep things simple, we're going to just close the app in these very
exceptional cases.