check if the identity key pair exists before creation
sometimes we create the identity key pair although it already exists
we try to avoid that by checking the group again right before sending the request.
we do not force reloading the cache as this is done outside by the caller already or the group is new. so it should not be necessary.