Fix computeColor throwing on non-srgb color spaces
This would cause some emails to not render in dark mode.

Basically, MDN claimed that all colors from getComputedStyle will be in
rgb or rgba form. However, they have since updated their page, and it
turns out the getComputedStyle ALSO passes through colors that are in
a given colorspace.

To fix this, we can transmute the color into srgb using the CSS color()
function and extract the RGB/A components using a regular expression.

Co-authored-by: hrb <hrb-hub@users.noreply.github.com>