fix aggregate_type_model using clientModel instead of serverModel
When decrypting an entity such as Mail on the rust SDK, we were wrongly
using the clientModel to resolve the aggregate_type_model of an
aggregated entity. This lead to an error when trying to decrypt an
entity that has a new aggregated type in a newer serverModel. The
entire decryption procedure should always use the serverModel.

Co-authored-by: abp <abp@tutao.de>
Co-authored-by: map <mpfau@users.noreply.github.com>