fix Android compileSdk version in tutasdk build.gradle.kts
The Android compileSdk version in the tutasdk build.gradle.kts has been
wrongly set to a different compileSdk version 34, instead of version 36
as with the other gradle artifacts (mail, calendar, shared). This
version discrepancy caused, when running the Android release build,
to not bundle the "libtutasdk.so" file correctly, leading to an
Android APK file without native rust code.

Co-authored-by: abp <abp@tutao.de>
Co-authored-by: das <das@tutao.de>