Simplify event invitation banner text
In order to avoid layout issues within the event invitation preview banner
we keep the length of the text to a minimum.

The idea is that the expander should only show "x conflicts" and doesn't show a
preview of which type of conflict it is. The concrete type of the conflict (all day or simultaneous)
can seen by expanding the conflict section.