use a smarter method for rate-limiting email availability
we have a rate limit on the server, but it's easy to run into and the
resulting suspension will cause a 30s wait.

This new method should limit the request rate (from one client) to just
under the limit the server has, without feeling slow for the vast
majority of users.