Fix null in associations for PatchMerger.ts
We are attempting to fix the case where a null value was added to the
association, it has happened that a mail instance in the offline
database ended up with null in the list ([null]) for the 1729
aggregation.

This happened on the parsedInstance level, and for a zeroOrOne
aggregation it should have been just an empty list.