read full instance from UPDATE entityUpdate if available
Usually, entityUpdates with operation UPDATE only have patches
on the entityUpdate instead of the full instance, like e.g. on
entityUpdates with operation CREATE. However, in certain
scenarios we are still / again sending full instances on entityUpdates
with operation UPDATE, and should therefore also correctly process them.
This commit change the processing of UPDATE entityUpdates to be:
-> use entityUpdate.patches
--> if not available use entityUpdate.instance
---> if not available re-load instance from server

This commit fixes issues where instance were re-loaded from the server
even though the instances was available on the entityUpdate.