add age confirmation to settings & news item
since we're removing the age checkbox from signup, we have
to add it to the usage test opt-in since we can't get consent
for that from someone that's 15 or younger (it has to be a legal
guardian).

We accomplish this by adding the checkbox as a required step to activate
usage tests from the news item and with a new dialog that asks for
confirmation after selecting "Activated" in the settings.