save spam classification model in IndexedDB on webApp
From now on, we save the trained SpamClassificationModel for each
mailGroup (mailbox) to the IndexedDB on the webApp. On devices where
OfflineStorage is available, we save it to the OfflineStorage. This
separation is handled by the SpamClassifierStorageFacade.

This commit furthermore fixes a bug, where the metadata, such
as lastTrainedFromScratchTime and lastTrainingDataIndexId were not saved
per mailGroup (mailbox), but instead only globally. We now bundle them
with the SpamClassificationModel as SpamClassificationModelMetaData.

When removing a group membership with GroupType.Mail the
CustomUserCacheHandler does now also remove any stored
SpamClassificationModel corresponding to this mailGroupId.

Co-authored-by: jomapp <17314077+jomapp@users.noreply.github.com>
Co-authored-by: das <das@tutao.de>