apply inbox rules for mails delivered to the spam folder
In case server delivers a mail to the spam folder, it was possible that
the client spam classifier classifies it as ham and moves it to the
inbox folder. In that case, those mails would end up staying in inbox
even if there are inbox rules suitable for them. This commit fixes that
and applies inbox rules to mails landing in spam.