use existing spam classification model for prediction while retraining
From now on, we initialize the existing spam classification model
already onPartialLoginSuccess, to be able to do predictions as early
as possible. Additionally, while retraining a model, we keep the old
model to allow predictions while retraining is still ongoing.
Re-training as well as full training now happens only after the sync
is done, as indicated by the SyncTracker.

Co-authored-by: abp <abp@tutao.de>