Don't have all mails loaded at once with SQLite search
We can have all of the IDs loaded at once since those are small, but
having all of the mail instances loaded at once can lead to OOM on low
RAM devices like phones.

Instead we can hold back all of the results and extend it as we need
more results.

Closes #9976