remove _finalIvs field from the instances
Previously, we needed to persist the initialization vectors for
encrypted final fields to make sure that we get the same ciphertext
when we encrypt them so that we could verify that these fields didn't
change during a resource update (PUT). Final fields are the fields that
cannot be modified by the client. As we do not do PUT requests for
resource updates and send only the changed fields with a PATCH request,
we don't need to persist the initialization vectors on the instance
anymore.

Co-authored-by: das <das@tutao.de>
Co-authored-by: jomapp <17314077+jomapp@users.noreply.github.com>