improve ham/spam ratio calculation in SpamClassifierDataDealer.ts
We improved the ham/spam ratio calculation to better support
case where a mailbox does not have any ham or spam mails. In case
a mailbox now has 0 ham mails or 0 spam mails we use the ratio
DEFAULT_HAM_COUNT/x or x/DEFAULT_SPAM_COUNT. This makes sure that we
do not train with all mails, but subsample correctly. Furthermore,
this commit increases the HIGH_CONFIDENCE_THRESHOLD to 4, which enables
separation between mails that have been moved by an inbox rule and
mails that have been moved manually by the user in the future.

Co-authored-by: abp <abp@tutao.de>
Co-authored-by: das <das@tutao.de>