Re-create event wrappers when changing calendar color.
With the factoring of the calendar time based views we introduced an EventWrapper around every event
that holds certain attributes of how to render event bubbles, including the calendar color.

When changing a color we need to update all existing event wrappers, so the appropriate calendar color gets updated.
fix #10071