hide mails in mail list when initial processing is not done yet
With the implementation of the new client-side spam classification
model, we introduced a new boolean flag processNeeded on the mail type.
This flag is set to true, iff a mail has not been processed by the
ProcessInboxHandler (SpamClassificationHandler and InboxRuleHandler),
and therefore spam processing and inbox rule application has not been
executed yet. In this state the folder a newly received mail
should be shown in is not final yet. With hiding mails where
processNeeded == true, a user does not see mails moving around
randomly in and out of the inbox or spam folder.
Important to note:
We explicitly do NOT remove mails with processNeeded == true from
the underlying listModel (ConversationListModel or MailListModel) to
always properly respect the "current state" of existing MailSetEntries
and Mails on the client before the entityUpdate with
processNeeded == false is received for the corresponding mail.

Co-authored-by: abp <abp@tutao.de>