Calculate number of rows for all days before calculating height of all day section
In the all day section we need to calculate the proper height so all event bubbles can fit into the section. In case there is a  change to the section, e.g. by hiding a calendar, we should just rely on the result of layoutEvents and not maintain a state inside the view.
fix #10089