enforce usage of v (model version) header on every (non-legacy) request
We can now enforce that every request must have the v header set, it
contains the model version corresponding to the instance transmitted.
Previously, the v header was used to check for deprecated clients,
but with the introduction of the cv header (full client version), this
is obsolete. From now on only the cv header is used to check
deprecations and throw the InvalidSoftwareVersionException.

Co-authored-by: das <das@tutao.de>