do not immediately create offline db for post-signup session
the post-signup session is a bit special in that it is the second
session created after (re-)loading the page. The first one is the
temporary session used to update the freshly created account with
payment info etc. In all other cases, logging out reloads the page and
clears all state.

Logging out of the temporary session seems to leave some state behind
that's not properly cleaned up that causes the post-signup session to
ignore some entity events or not store some instances in its persistent
offline DB under certain circumstances. This causes login to fail after
reloading the page.

Current working theory is that the offline DB does not contain some
instances it should contain when it processes the first entity events or
that the event bus or another cache maintains some state after logout
that leads to an inconsistency in the offline DB.

The fix is to not create a persistent offline DB in
createPostSignupSession, but storing a database key. This will cause a
fresh offlineDB to be initialized on the first login, sidestepping the
issue.

close https://github.com/tutao/tutanota/issues/10079