propagate decryption errors in nested aggregates to the parent instance
previously errors in nested aggregates were not visible at the top instance level. in case there is not value at the top level that fails decryption (which is the case for example for MailDetailsBlob) these instances would not be treated like errors. this leads for example to the case that we would cache such instances and then never properly retry decryption although this would help with temporary errors.

now we bubble the _errors up to the top level of the instance, which allows for an easy check.

we use this check in case of MailDetails in MailViewerViewModel to decide how to render an instance

also we now properly error out on patches that contain decryption _errors