do not persist temporary TutaCrypt authentication errors
so far in case we got a temporary error when resolving the sender's identity key or public encryption key we would persist the error and throw away the bucket key.

this made such an error indistinguishable from a real MITM attack.

we now fail decryption completely in those temporary error cases and will retry later. this should lead to less unnecessary user warnings that would otherwise render the key verification feature unreliable in the long run.