add in-memory cache for public encryption keys
we must reduce the number of requests to the PublicKeyService so we cache the result after verification
we do not cache the verification result as we might later verify a key and want to benefit from that immediately
for the same reason we only cache signed keys. otherwise we could not verify in case our counterpart signs it after it was cached

the one exception from that rule is the system user group (the sender of external mails). this group does not have an identity key (and we also bypass the IdentityKeyService request in this case) but we still want to cache here, because this key should never change and verifying a future potential signature would anyway require to code change to fetch the respective identity key.