Add dv (dependsOnVersion) header to requests for types with dependency
We added the dependsOnVersion field to the TypeModel, which is set only
for the clientTypeModel. It is used to inform the server of the external
dependency version for the instance being transmitted with the
rest request, only the instanceType has a non-null dependency.

Every application (e.g. tutanota, sys, etc.) can only
depend on one other application. We exclusively use this header to
forward migrate external aggregated entities on the server, received
by either POST or PATCH requests.

Co-authored-by: jhm <17314077+jomapp@users.noreply.github.com>
Co-authored-by: sug <sug@tutao.de>